Create seed data
Future goal will be to create Requirement/Event/Participation tables in
such a way that the requirements can handle multiple organizations.
There's nothing linking a requirement to a specific group right now,
which is fine for the first sprint, but essentially makes the site
applicable to only one group right now.