Construct Base UI
The mocked UI has been constructed using basic bootstrap themes. A sidebar for navigation as well as a list appear in main/index. The list is the contents of requirements from the controller. This completes Task 3 of the Application Design story in Sprint 1.

The next step would be to provide navigation controls which allow the user to navigate through different sets of requirements as specified in the database.