Fix Issue Number 5
The width of the progress circle (.circularProgressBorder) has been upped to 110px instead of 92px. A dynamic sizing approach was considered, however that would cause negative effects in the visual hierarchy of the user interface due to different circle sizes between rows.

The solution of upping to 110px was chosen to keep all circles uniform and able to contain three digits in the numerator.