Fix Completion Progress Storage
Completion progress was being stored in the requirements table, however this table should be generic for use by all organizations at a given school. The completion progress was moved from the requirements table and placed in events. This way when an event is added to the system, the processing controller can calculate the amount of completion that the given event will yield, and save that to the database.

The main views have been updated to show all requirements as fully completed, the logic still needs to be added to compute actual completion.