Provide Tier Status to User
The main controller now calculates the current tier expected for the group based on their progress towards requirements. An overview view shows the expected tier and the points until the next tier.

The numbering scheme for requirements has also been updated to correspond to the actual requirement section and number.