Event Summation fixed for issue #8
"requirement_progress" is not in the database table thus an error
retrieving the value.

A check if value is above bounds so that the progress bar displays green
for factions greater than 1 (e.g. 10/5)

Added more values to seeds (to cover some in achedemics) more values
will need to be added for the other sections